# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Petstore - OpenAPI 3.0",
    "description": "UCL winners",
    "contact": {
      "email": "luka.balic@fer.hr"
    },
    "license": {
      "name": "Creative Commons Zero v1.0 Universal"
    },
    "version": "1.0.11"
  },
  "paths": {
    "/coach/all": {
      "get": {
        "tags": [
          "coach"
        ],
        "summary": "Fetch all coaches",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/coach/{Id}": {
      "get": {
        "tags": [
          "coach"
        ],
        "summary": "Fetch coach by ID",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "ID of coach to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "Coach not found"
          }
        }
      }
    },
    "/player/all": {
      "get": {
        "tags": [
          "player"
        ],
        "summary": "Fetch all players",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/player/{Id}": {
      "get": {
        "tags": [
          "player"
        ],
        "summary": "Fetch player by ID",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "ID of player to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "Player not found"
          }
        }
      }
    },
    "/uclWinner/all": {
      "get": {
        "tags": [
          "uclWinner"
        ],
        "summary": "Fetch all UCL winners",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/uclWinner/{Id}": {
      "get": {
        "tags": [
          "uclWinner"
        ],
        "summary": "Fetch uclWinner by ID",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "ID of an UCL winner to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "Coach not found"
          }
        }
      }
    },
    "/uclWinner/update": {
      "put": {
        "tags": [
          "uclWinner"
        ],
        "summary": "Updates an UCL winner",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Id of the UCL winner that needs to be updated",
            "schema": {
              "type": "integer"
            },
            "required": true
          },
          {
            "name": "club",
            "in": "query",
            "description": "Club name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "in": "query",
            "description": "City name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "country",
            "in": "query",
            "description": "Country name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Year",
            "in": "query",
            "description": "year",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "coach",
            "in": "query",
            "description": "Coach",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "captain",
            "in": "query",
            "description": "Captain",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "result",
            "in": "query",
            "description": "Result",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "attendance",
            "in": "query",
            "description": "Attendance",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stadium",
            "in": "query",
            "description": "Stadium",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "matchup",
            "in": "query",
            "description": "Matchup",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "UCL winner updated"
          },
          "404": {
            "description": "UCL winner not found"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/uclWinner/add": {
      "post": {
        "tags": [
          "uclWinner"
        ],
        "summary": "Adds an UCL winner",
        "description": "Add an UCL winner",
        "parameters": [
          {
            "name": "club",
            "in": "query",
            "description": "Club name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "in": "query",
            "description": "City name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "country",
            "in": "query",
            "description": "Country name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Year",
            "in": "query",
            "description": "year",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "coach",
            "in": "query",
            "description": "Coach",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "captain",
            "in": "query",
            "description": "Captain",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "result",
            "in": "query",
            "description": "Result",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "attendance",
            "in": "query",
            "description": "Attendance",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stadium",
            "in": "query",
            "description": "Stadium",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "matchup",
            "in": "query",
            "description": "Matchup",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "UCL winner added"
          },
          "400": {
            "description": "UCL winner alredy exists"
          }
        }
      }
    },
    "/uclWinner/delete/{Id}": {
      "delete": {
        "tags": [
          "uclWinner"
        ],
        "summary": "Deletes an UCL winner",
        "description": "delete an UCL winner",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "UCL winner deleted"
          },
          "404": {
            "description": "UCL winner does not exist"
          }
        }
      }
    }
  }
}